+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## next
 
+## `@svizzle/barchart` v0.5.0 (next)
+
+- add prop `titleFontSize`
+
 ## `@svizzle/choropleth` v0.4.0 (next)
 
 - add `projectionId` so that we can pass either a `projection` prop (a projection function) or an id to select a projection from the available ones